requirements.txt
# Core audio + metadata tooling
mutagen
pydub
simpleaudio
ffmpeg-python
pyacoustid
musicbrainzngs
python-vlc

# Audio analysis + clustering
librosa
scipy
numpy
scikit-learn
hdbscan
matplotlib
Pillow

# HTTP + typing helpers
requests
typing_extensions

# Optional metadata helpers
spotipy

# Optional UX helpers
tqdm
ttkthemes
ttkbootstrap
sv-ttk
PySide6
pyqtgraph>=0.13.0

# Optional alternative audio engine; see README.md for install steps
# essentia==2.1b6
